Adds NginxConfigurator
To facilitate custom nginx configs and install locations, this uses nginx's configuration
settings to detect the conf directory, add the vhosts' 'sites' directory, and load to
Nginx in a more OO-oriented way.  This allows swapping out Nginx for Apache or even a bare
Ruby reverse proxy system in the future.